requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.3
Pillow==11.2.1
streamlit==1.45.0
streamlit_option_menu==0.4.0
